ecommerce-kafka-elk/producer/requirements.txt
kafka-python==2.0.2
pandas==1.5.3
numpy==1.23.5
python-dotenv==1.0.0